Chore - do some refactoring for the project validation
