+++ packages/toolkit/src/amazonqGumby/activation.ts
@@ -13,16 +13,11 @@ import { transformByQState } from '../codewhisperer/models/model'
 import * as CodeWhispererConstants from '../codewhisperer/models/constants'
 import { ProposedTransformationExplorer } from '../codewhisperer/service/transformationResultsViewProvider'
 import { codeTransformTelemetryState } from './telemetry/codeTransformTelemetryState'
-import { CodeTransformJavaSourceVersionsAllowed, telemetry } from '../shared/telemetry/telemetry'
-import {
-    CancelActionPositions,
-    JDKToTelemetryValue,
-    logCodeTransformInitiatedMetric,
-} from './telemetry/codeTransformTelemetry'
+import { telemetry } from '../shared/telemetry/telemetry'
+import { CancelActionPositions, logCodeTransformInitiatedMetric } from './telemetry/codeTransformTelemetry'
 import { CodeTransformConstants } from './models/constants'
-import { getOpenProjects, validateOpenProjects } from '../codewhisperer/service/transformByQHandler'
-import { MetadataResult } from '../shared/telemetry/telemetryClient'
 import { AuthUtil } from '../codewhisperer/util/authUtil'
+import { validateAndLogProjectDetails } from '../codewhisperer/service/transformByQHandler'
 
 export async function activate(context: ExtContext) {
     // If the user is codewhisperer eligible, run these checks on startup
@@ -90,24 +85,6 @@ export async function activate(context: ExtContext) {
             workspaceChangeEvent
         )
 
-        // Try to validate project silently
-        try {
-            const openProjects = await getOpenProjects()
-            const validProjects = await validateOpenProjects(openProjects)
-            const firstKey = validProjects.keys().next().value
-            const firstModuleEntry = validProjects.get(firstKey)
-            telemetry.codeTransform_projectDetails.emit({
-                codeTransformSessionId: codeTransformTelemetryState.getSessionId(),
-                codeTransformLocalJavaVersion:
-                    (JDKToTelemetryValue(firstModuleEntry) as CodeTransformJavaSourceVersionsAllowed) || undefined,
-            })
-        } catch (err: any) {
-            telemetry.codeTransform_projectDetails.emit({
-                codeTransformSessionId: codeTransformTelemetryState.getSessionId(),
-                codeTransformPreValidationError: err?.name || undefined,
-                result: MetadataResult.Fail,
-                reason: err?.code || undefined,
-            })
-        }
+        validateAndLogProjectDetails()
     }
 }
